Move all string functions from all Windows projects to a standalone lib.
We can't rely on the Win32 API for character set conversions, because Windows NT 4.0 does not support UTF-8.
We can even rely less on std::locale to do the same, because that doesn't even work with Windows 2000.

The new conversion functions should be faster and smaller at the same time.