Compile against EnlyzeWinCompatLib's libc++ and winpthreads.
Otherwise, it compiles against MSVC's C++ standard library, resulting in compatibility problems.

This enforces a certain project directory structure, but we currently use that for all our Windows software.
It also makes a standalone build of EnlyzeWinStringLib impossible, but we don't need that anyway.
Remove the .sln to indicate that.